# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PokéVault API — private Pokémon card collection and resale tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve a private object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  # Cards
...
  /trash:
    get:
      tags: [Trash]
      operationId: listTrash
      summary: List soft-deleted cards and sealed products
      responses:
        '200':
          description: Trashed items sorted by deletion date.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrashItem'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /trash/cards/{id}/restore:
    post:
      tags: [Trash]
      operationId: restoreCard
      summary: Restore a soft-deleted card from trash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Restored card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /trash/sealed-products/{id}/restore:
    post:
      tags: [Trash]
      operationId: restoreSealedProduct
      summary: Restore a soft-deleted sealed product from trash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Restored sealed product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SealedProduct'
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /trash/cards/{id}:
    delete:
      tags: [Trash]
      operationId: hardDeleteCard
      summary: Permanently delete a trashed card and its images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Permanently deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /trash/sealed-products/{id}:
    delete:
      tags: [Trash]
      operationId: hardDeleteSealedProduct
      summary: Permanently delete a trashed sealed product and its images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Permanently deleted.
        '404':
          description: Not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Auth
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    # Images
...
    Card:
      type: object
      required: [id, name, set, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        set:
          type: string
        cardNumber:
          type: [string, 'null']
        rarity:
          type: [string, 'null']
        condition:
          type: [string, 'null']
        grade:
          type: [string, 'null']
        grader:
          type: [string, 'null']
        quantity:
          type: integer
        purchasePrice:
          type: [number, 'null']
        marketValue:
          type: [number, 'null']
        askingPrice:
          type: [number, 'null']
        soldPrice:
          type: [number, 'null']
        soldDate:
          type: [string, 'null']
          format: date
        status:
          type: string
          enum: [collection, for_sale, sold, wishlist, trade_binder, grading_pile]
        purchaseDate:
          type: [string, 'null']
          format: date
        purchaseSource:
          type: [string, 'null']
        tradeNotes:
          type: [string, 'null']
        location:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        coverImagePath:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, 'null']
          format: date-time
...
    SealedProduct:
      type: object
      required: [id, name, productType, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        productType:
          type: string
          enum: [ETB, booster_box, booster_pack, tin, binder, case, other]
        set:
          type: [string, 'null']
        quantity:
          type: integer
        condition:
          type: [string, 'null']
        purchasePrice:
          type: [number, 'null']
        marketValue:
          type: [number, 'null']
        askingPrice:
          type: [number, 'null']
        soldPrice:
          type: [number, 'null']
        soldDate:
          type: [string, 'null']
          format: date
        status:
          type: string
          enum: [collection, for_sale, sold, wishlist, trade_binder, grading_pile]
        purchaseDate:
          type: [string, 'null']
          format: date
        purchaseSource:
          type: [string, 'null']
        tradeNotes:
          type: [string, 'null']
        location:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        coverImagePath:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, 'null']
          format: date-time
...
    TrashItem:
      type: object
      required: [id, name, type, deletedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [card, sealed_product]
        status:
          type: string
        coverImagePath:
          type: [string, 'null']
        deletedAt:
          type: string
          format: date-time
...